bug 1079436 - fix validThrough as returned by VerifyEncodedOCSPResponse r=briansmith
validThrough should now be the time through which, if passed in as the given
time to validate an OCSP response at, VerifyEncodedOCSPResponse will still
consider it trustworthy. After that time, it will be expired. This makes it
so the OCSP cache compares validity period responses consistently with
mozilla::pkix.

BUG=https://bugzilla.mozilla.org/show_bug.cgi?id=1079436

Imported-from: https://hg.mozilla.org/mozilla-central/rev/b379f1bc58e1