Bug 1154399 - Part 3: Simplify OptionalExtensions. r=keeler
We used to avoid using Nested and NestedOf because they were based on
bind and it was difficult to maintain our std::bind polyfill. Now that
we use lambdas, it is easy to use Nested and NestedOf, so we should do
so wherever it makes the code clearer.

BUG=https://bugzilla.mozilla.org/show_bug.cgi?id=1154399

Imported-from: https://hg.mozilla.org/mozilla-central/rev/37975b5017c4