Add adapter for OpenSSL's libcrypto.
1. In lib/pkixlibcrypto.cpp, add adapters that simplify the creation of
mozilla::pkix TrustDomain implementations based on OpenSSL's libcrypto.
This commit adds support for signature verification and digest
calculation. In a future commit, we'll add support for using
OpenSSL's certificate trust store. lib/pkixcrypto.cpp is the OpenSSL
counterpart to the NSS adapter in lib/pkixnss.cpp.

2. In test/lib/pkixtestlibcrypto.cpp, implement the test crypto
interface so that the unit tests can be run using OpenSSL's libcrypto.
test/lib/pkixtestlibcrypto.cpp is the OpenSSL counterpart to the NSS
implementation of the test interface in test/lib/pkixtestnss.cpp. Note
that the DSS and MD2 portions of the test interface are NOT implemented
yet.