Simplify TrustDomain signature verification interface.
The original plan was for mozilla::pkix to do all the parsing of the
internal structures of public keys and ECDSA signatures itself,
bypassing any of the parsing done by the crypto library. However, that
would have meant that mozilla::pkix's TrustDomain adapters would have
had to add dependencies on low-level details of the crypto
implementation such as the bugnum API. That would have been especially
bad because those are the details that are most likely to change in
incompatible ways in OpenSSL, BoringSSL, and ring. The parsing of these
encoded structures in BoringSSL is much more robust than before, so
mozilla::pkix usually wouldn't be buying much additional safety by
doing that parsing itself.